pyproject.toml
[tool.poetry]
name = "medium-clone"
version = "0.1.0"
description = ""
authors = ["morveusz93 <morveusz93@gmail.com>"]
readme = "README.md"
packages = [{include = "medium_clone"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.3"
django-tinymce = "^3.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
